Fix schemas
